Add First voter Courses + Simple 403 page

+++ src/Controller/Back/CourseController.php
@@ -31,6 +31,9 @@ class CourseController extends AbstractController
     public function new(Request $request): Response
     {
         $course = new Course();
+
+        $this->denyAccessUnlessGranted('create', $course);
+
         $form = $this->createForm(CourseType::class, $course);
         $form->handleRequest($request);
 
@@ -63,6 +66,8 @@ class CourseController extends AbstractController
      */
     public function edit(Request $request, Course $course): Response
     {
+        $this->denyAccessUnlessGranted('edit', $course);
+
         $form = $this->createForm(CourseType::class, $course);
         $form->handleRequest($request);
 
@@ -85,6 +90,8 @@ class CourseController extends AbstractController
      */
     public function delete(Request $request, Course $course): Response
     {
+        $this->denyAccessUnlessGranted('delete', $course);
+
         if ($this->isCsrfTokenValid('delete'.$course->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($course);
